UserInterface: Updated to use mContentResolver
AOKP source uses mContentResolver in place of mContext.getContentResolver() for calling system settings, this was causing build errors whenever I would cherry-pick and forget to then modify any references.

I found it easier to just modify the UserInterface fragment to this style so future cherry-picks would not cause this issue.

Change-Id: Ida5dab47b6240cf9eb4cb750d87b11766100e355